move Not type to constraint.operators package
